The official GDAL VRT XML schema A copy of the official VRT schema for vritility VRT validation.

Loads a copy of the official xml schema from the provided vrt_xml_schema dataset.

Usage

vrt_xml_schema

vrt_schema(schema = vrtility::vrt_xml_schema)

Format

character

Source

https://raw.githubusercontent.com/OSGeo/gdal/master/frmts/vrt/data/gdalvrt.xsd

Arguments

schema

a character string of the xml schema - mainly intended only for use with the official gdal VRT schema.

Value

An xml_document object

Examples

vrt_schema()
#> {xml_document}
#> <schema elementFormDefault="qualified" version="1.0" xmlns:xs="http://www.w3.org/2001/XMLSchema">
#>  [1] <xs:element name="VRTDataset" type="VRTDatasetType">\n  <xs:annotation>\ ...
#>  [2] <xs:complexType name="VRTDatasetType">\n  <xs:sequence>\n    <xs:choice  ...
#>  [3] <xs:simpleType name="DatasetSubclassType">\n  <xs:restriction base="xs:s ...
#>  [4] <xs:complexType name="OverviewListType">\n  <xs:simpleContent>\n    <xs: ...
#>  [5] <xs:simpleType name="integerList">\n  <xs:list itemType="xs:integer"/>\n ...
#>  [6] <xs:complexType name="SRSType">\n  <xs:simpleContent>\n    <xs:extension ...
#>  [7] <xs:simpleType name="nonNegativeInteger32">\n  <xs:restriction base="xs: ...
#>  [8] <xs:complexType name="GCPListType">\n  <xs:sequence>\n    <xs:element na ...
#>  [9] <xs:complexType name="GCPType">\n  <xs:attribute name="Id" type="xs:stri ...
#> [10] <xs:complexType name="MetadataType">\n  <xs:sequence>\n    <!--<xs:choic ...
#> [11] <xs:complexType name="GDALWarpOptionsType">\n  <xs:sequence>\n    <xs:an ...
#> [12] <xs:complexType name="PansharpeningOptionsType">\n  <xs:sequence>\n    < ...
#> [13] <xs:simpleType name="NoDataOrNoneType">\n  <xs:union memberTypes="xs:dou ...
#> [14] <xs:complexType name="PanchroBandType">\n  <xs:sequence>\n    <xs:elemen ...
#> [15] <xs:complexType name="SpectralBandType">\n  <xs:sequence>\n    <xs:eleme ...
#> [16] <xs:complexType name="AlgorithmOptionsType">\n  <xs:sequence>\n    <xs:a ...
#> [17] <xs:complexType name="InputType">\n  <xs:sequence>\n    <xs:choice minOc ...
#> [18] <xs:complexType name="OutputBandsType">\n  <xs:sequence/>\n  <xs:attribu ...
#> [19] <xs:complexType name="ProcessingStepsType">\n  <xs:sequence minOccurs="1 ...
#> [20] <xs:complexType name="ProcessingStepType">\n  <xs:annotation>\n    <xs:d ...
#> ...