vrtility: GDAL VRT utilities for R

vrtility is a package for building raster (primarily remote sensing) processing pipelines. It makes use of GDAL's VRT (virtual raster format) capabilities for efficient processing of large raster datasets. This package's primary focus is on the use of GDAL VRT pixel functions. These pixel functions (currently implemented with python) are used to apply cloud masks and summarise pixel values (e.g. median) from multiple images (i.e create a composite image). We hope to add C++ or expression based pixel functions in time.

Spatial helpers

bbox_to_projected

Convert a long/lat bounding box to a projected bounding box

to_wkt

Get a wkt string from a numeric epsg code or character such as a proj4 string

STAC helpers

stac_query

Query a STAC catalog

sentinel2_stac_query

Query a STAC catalog for Sentinel-2 data

sign_planetary_computer

Sign a URL for the Planetary Computer

VRT utilities

vrt_collect

Create a vrt_collection object

vrt_set_maskfun

Set a mask function for a vrt object

vrt_stack

Create a vrt_stack object

vrt_set_py_pixelfun

Set a pixel function for a vrt stack object

vrt_set_gdal_pixelfun

Set a GDAL pixel function for a vrt stack object

vrt_set_scale

Set the scale and offset values for a vrt_x object

vrt_move_band

Move a band in a vrt_x object

vrt_set_nodata

Set the nodata value for a vrt_x object

vrt_warp

Warp a vrt_x object to a warped vrt

vrt_compute

Compute a vrt pipeline (using GDAL)

VRT pixel functions

vrt_set_py_pixelfun

Set a pixel function for a vrt stack object

set_mask_numpy

A pixel function to apply a bitmask

build_intmask

A pixel function to build a mask from a mask band where the mask values are integers

build_bitmask

A pixel function to build a mask from a mask band where the mask values are bit positions

median_numpy

A pixel function to compute the median

python environment helpers

vrtility_py_require

Require a python package

set_py_env_vals

Set python environment variables as options

VRT helpers

vrt_save

Save a VRT object to a file

vrt_schema

(data object)The official GDAL VRT schema as a character object

See also

Author

Maintainer: Hugh Graham hugh.graham@permianglobal.com (ORCID)